cmp - integer cast warning

+++ SRC/material/section/FiberSection3dThermal.cpp
@@ -981,7 +981,7 @@ FiberSection3dThermal::Print(OPS_Stream &s, int flag)
   else {
     s << "\nFiberSection3dThermal, tag: " << this->getTag() << "\n";
     s << "\tSection code: " << code;
-    s << "\tNumber of Fibers: " << fibers.size() << "\n";
+    s << "\tNumber of Fibers: " << static_cast<int>(fibers.size()) << "\n";
     s << "\tCentroid: (" << yBar << ", " << zBar << ')' << "\n";
     if (theTorsion != 0)
         theTorsion->Print(s, flag);
